# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: SupportAI API
  version: 0.1.0
servers:
  - url: http://localhost:4000
paths:
  /health:
    get:
      summary: Health check
      responses:
        "200":
          description: API is healthy
  /api/v1/profile:
    get:
      summary: Get authenticated user profile and memberships
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Profile returned
        "401":
          description: Authentication required
  /api/v1/businesses:
    post:
      summary: Create business tenant
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, industry]
              properties:
                name:
                  type: string
                industry:
                  type: string
      responses:
        "201":
          description: Business created
  /api/v1/documents:
    get:
      summary: List documents
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Documents returned
    post:
      summary: Upload and process a document
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "201":
          description: Document uploaded
  /api/v1/documents/{documentId}:
    delete:
      summary: Delete document
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: documentId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Deleted
  /api/v1/faqs:
    get:
      summary: List FAQs
      security:
        - bearerAuth: []
      responses:
        "200":
          description: FAQs returned
    post:
      summary: Create FAQ
      security:
        - bearerAuth: []
      responses:
        "201":
          description: FAQ created
  /api/v1/faqs/{faqId}:
    patch:
      summary: Update FAQ
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: faqId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: FAQ updated
    delete:
      summary: Delete FAQ
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: faqId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Deleted
  /api/v1/chat/conversations:
    get:
      summary: List conversations
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Conversations returned
    post:
      summary: Create conversation
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Conversation created
  /api/v1/chat/conversations/{conversationId}:
    get:
      summary: Get conversation with messages
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: conversationId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Conversation returned
  /api/v1/chat/conversations/{conversationId}/messages:
    post:
      summary: Ask a question and generate a RAG answer
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: conversationId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "201":
          description: Answer generated
  /api/v1/analytics/summary:
    get:
      summary: Get dashboard analytics
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Analytics returned
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Supabase JWT